db/group: add initial group db helpers
